fix: move page into [lang]
