+++ app/[lang]/notifications.tsx
